baseline/하은/requirements.txt
streamlit>=1.44.0
olefile>=0.47
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
pandas>=2.2.0
pyhwp>=0.1b15
langchain-text-splitters>=0.3.11
rank-bm25>=0.2.2
python-docx>=1.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
chromadb>=1.5.0
openai>=2.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
